+++ tests/e2e/ui/tooltip/app.py
@@ -4,8 +4,6 @@ app_ui = ui.page_fluid(
     ui.input_action_button("btn_show", "Show tooltip", class_="mt-3 me-3"),
     ui.input_action_button("btn_close", "Close tooltip", class_="mt-3 me-3"),
     ui.br(),
-    ui.input_action_button("btn_toggle", "Toggle tooltip", class_="mt-3 me-3"),
-    ui.br(),
     ui.br(),
     ui.tooltip(
         ui.input_action_button("btn_w_tooltip", "A button w/ a tooltip", class_="mt-3"),
@@ -21,19 +19,13 @@ def server(input: Inputs, output: Outputs, session: Session):
     def _():
         req(input.btn_show())
 
-        ui.toggle_tooltip("tooltip_id", show=True)
+        ui.update_tooltip("tooltip_id", show=True)
 
     @reactive.Effect
     def _():
         req(input.btn_close())
 
-        ui.toggle_tooltip("tooltip_id", show=False)
-
-    @reactive.Effect
-    def _():
-        req(input.btn_toggle())
-
-        ui.toggle_tooltip("tooltip_id")
+        ui.update_tooltip("tooltip_id", show=False)
 
     @reactive.Effect
     def _():